Add ability to measure distances
The anchor application code is now updated to match the tag
application code so actual measurements takes place.

In this commit, up to 3 anchors can be used with 1 tag to measure
distances between each anchor/tag pair. In this very rudimentary
measurement, the tag can be localised using the 3 distances measured.

However, the time stamps are not well calibrated yet. Hence the
distances measured are not accurate nor precise enough yet. Further
calibration through the manipulation of the delay values used as well
as improving the ranging technique can be used to achieved better
results.