Add ability to dynamically compute delay
The delay for anchors to respond to tag's initiation message are now
dynamically computed by using a 1-byte number that specifies the
total number of anchors currently in this exchange system.

This number is set by the tag and sent out in the initiation message
where each anchor will read this value embedded in the message.