Fix Anchor failing to poll for correct message
A fix has been attempted to synchronise message exchange between the
Tag and Anchor. More specifically, when Anchor needs to poll for
Tag's final message.

Previously, Anchor will poll for final message repeatedly, ONCE ONLY.
What this means is that if the Anchor receives a frame but it is not
the final message, the application continues as if nothing happens.
The fix is to add a while loop condition check to ensure we will
always poll for final message even after receiving different frame.