Add reception of data from RX buffer
Ability to receive and read date from RX buffer of the nRF52832 chip.

Up to 20 bytes of data could be read from the RX buffer. The data is
read into an array of unsigned chars where the Tag application will
decode the data according to a common data format.

The data format will have to be thoughtfully constructed, WIP.