Fix UART reception
The 'app_uart_evt_t' event type 'APP_UART_DATA_READY' is triggered
for EVERY BYTE that is moved into the buffer.

Hence, we use a character (';' in this case) to represent the end of
a transmission message. For every byte of data received, we check if
this byte is the end character. If it is, we know the message has
ended and so we notify the main application of incoming command.
Otherwise, we continue waiting for data in the buffer as we build up
the command message.