Refactor application
Apply Separation of Concerns to various sections of the code.
Separated 2 main sections of code, each with different purpose.

First is the message_processor. This section deals with write and
reading of messages that is to be transmitted and received at various
points of time.

Second is the timerstamper. This is simply functions to retrieve the
timestamp values when a transmission or reception occurs.