Fix late turning on of receiver
Issue:
At the start of every cycle, all nodes other than 'master' node 0
turns on their receiver to look for node 0's TX. However, due to
delay in the turning on of receiver, the nodes will miss node 0's
first transmission.

Fix:
Create a new timer (wake timer) along with sleep timer to ensure
the receivers are turned on properly and ready for reception from
node 0.
Wake timer is the time duration from the last TX/RX to the start of
the next cycle.
Sleep timer is the time duration of the actual hardware sleep
duration. Sleep timer value must be lesser than wake timer with a
large enough difference such that there is enough time for the
receiver to turn on and be ready for reception of frames.