Change the way Chunk ID is generated
A random number is used instead of the checksum of the payload data.

The reason being that if two Chunks have the same exact payload data,
the IDs of both Chunks will be the same. Hence having the same ID can
cause Segments transporting in the network to be handled incorrectly.